+++ src/shared/ProgressBar.h
@@ -24,7 +24,7 @@
 class barGoLink
 {
     public: // constructors
-        barGoLink(int row_count);
+        barGoLink(uint64 row_count);
         ~barGoLink();
 
     public: // modifiers
@@ -36,10 +36,10 @@ class barGoLink
         static char const * const empty;
         static char const * const full;
 
-        int rec_no;
-        int rec_pos;
-        int num_rec;
-        int indic_len;
+        uint64 rec_no;
+        uint64 rec_pos;
+        uint64 num_rec;
+        uint64 indic_len;
 };
 #endif
